# [Kernel][Performance] Tweak MoE Batched silu_mul_fp8_quant_deep_gemm kernel